Port MakeLocalFunctionStatic analyzer and tests to shared layer
CodeFix has not yet been ported because it depends on internal CodeGenerationSymbolFactory APIs. https://github.com/dotnet/roslyn/issues/43056 tracks making these APIs public and subsequently porting the fixer. To enable porting the tests, I have added a workaround in CodeStyle layer test framework to use the IDE code fixer with the CodeStyle layer analyzer for tests.
